requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
langchain>=0.3.14
langchain-openai>=0.3.0
langchain-core>=0.3.29
python-dotenv>=1.0.1
pydantic>=2.10.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
httpx>=0.27.0
redis>=5.0.0
email-validator>=2.0.0
greenlet>=3.0.0